+++ app/services/dbconnect.server.ts
@@ -1,7 +1,5 @@
 import mongoose from "mongoose";
-import { Users } from "../models/users.server";
 
-mongoose.connect("mongodb://localhost:27017/aimento");
-export let loader: any = async () => {
-  return await Users.find({});
-};
+export const connectServer = mongoose.connect(
+  "mongodb://localhost:27017/aimento"
+);
